// generated/mysqlx_session_pb.ts
export const file_mysqlx_session: GenFile /*@__PURE__*/ = fileDesc(
  "ChRteXNxbHhfc2Vzc2lvbi5wcm90bxIOTXlzcWx4LlNlc3Npb24iWQoRQXV0aGVudGljYXRlU3RhcnQSEQoJbWVjaF9uYW1lGAEgAigJEhEKCWF1dGhfZGF0YRgCIAEoDBIYChBpbml0aWFsX3Jlc3BvbnNlGAMgASgMOgSI6jAEIjMKFEF1dGhlbnRpY2F0ZUNvbnRpbnVlEhEKCWF1dGhfZGF0YRgBIAIoDDoIiOowBZDqMAMiKQoOQXV0aGVudGljYXRlT2sSEQoJYXV0aF9kYXRhGAEgASgMOgSQ6jAEIicKBVJlc2V0EhgKCWtlZXBfb3BlbhgBIAEoCDoFZmFsc2U6BIjqMAYiDQoFQ2xvc2U6BIjqMAdCGQoXY29tLm15c3FsLmNqLngucHJvdG9idWY",
  [file_mysqlx],
);

/**
 * *
 * The initial message send from the client to the server to start
 * the authentication process.
 * @returns @ref Mysqlx::Session::AuthenticateContinue
 *
 * @generated from message Mysqlx.Session.AuthenticateStart
 */
export type AuthenticateStart = Message<"Mysqlx.Session.AuthenticateStart"> & {
  /**
   * * authentication mechanism name
   *
   * @generated from field: required string mech_name = 1;
   */
  mechName: string;

  /**
   * * authentication data
   *
   * @generated from field: optional bytes auth_data = 2;
   */
  authData: Uint8Array;

  /**
   * * initial response
   *
   * @generated from field: optional bytes initial_response = 3;
   */
  initialResponse: Uint8Array;
};

/**
 * Describes the message Mysqlx.Session.AuthenticateStart.
 * Use `create(AuthenticateStartSchema)` to create a new message.
 */
export const AuthenticateStartSchema: GenMessage<AuthenticateStart> /*@__PURE__*/ = messageDesc(
  file_mysqlx_session,
  0,
);

/**
 * *
 * Send by client or server after an @ref Mysqlx::Session::AuthenticateStart
 * to exchange more authentication data.
 * @returns Mysqlx::Session::AuthenticateContinue
 *
 * @generated from message Mysqlx.Session.AuthenticateContinue
 */
export type AuthenticateContinue = Message<"Mysqlx.Session.AuthenticateContinue"> & {
  /**
   * * authentication data
   *
   * @generated from field: required bytes auth_data = 1;
   */
  authData: Uint8Array;
};

/**
 * Describes the message Mysqlx.Session.AuthenticateContinue.
 * Use `create(AuthenticateContinueSchema)` to create a new message.
 */
export const AuthenticateContinueSchema: GenMessage<AuthenticateContinue> /*@__PURE__*/ =
  messageDesc(file_mysqlx_session, 1);

/**
 * *
 * Sent by the server after successful authentication.
 *
 * @generated from message Mysqlx.Session.AuthenticateOk
 */
export type AuthenticateOk = Message<"Mysqlx.Session.AuthenticateOk"> & {
  /**
   * * authentication data
   *
   * @generated from field: optional bytes auth_data = 1;
   */
  authData: Uint8Array;
};

/**
 * Describes the message Mysqlx.Session.AuthenticateOk.
 * Use `create(AuthenticateOkSchema)` to create a new message.
 */
export const AuthenticateOkSchema: GenMessage<AuthenticateOk> /*@__PURE__*/ = messageDesc(
  file_mysqlx_session,
  2,
);
// ...
export type Close = Message<"Mysqlx.Session.Close"> & {};

/**
 * Describes the message Mysqlx.Session.Close.
 * Use `create(CloseSchema)` to create a new message.
 */
export const CloseSchema: GenMessage<Close> /*@__PURE__*/ = messageDesc(file_mysqlx_session, 4);