// generated/mysqlx_sql_pb.ts
export const file_mysqlx_sql: GenFile /*@__PURE__*/ = fileDesc(
  "ChBteXNxbHhfc3FsLnByb3RvEgpNeXNxbHguU3FsIn8KC1N0bXRFeGVjdXRlEhYKCW5hbWVzcGFjZRgDIAEoCToDc3FsEgwKBHN0bXQYASACKAwSIwoEYXJncxgCIAMoCzIVLk15c3FseC5EYXRhdHlwZXMuQW55Eh8KEGNvbXBhY3RfbWV0YWRhdGEYBCABKAg6BWZhbHNlOgSI6jAMIhUKDVN0bXRFeGVjdXRlT2s6BJDqMBFCGQoXY29tLm15c3FsLmNqLngucHJvdG9idWY",
  [file_mysqlx, file_mysqlx_datatypes],
);
// ...
export type StmtExecute = Message<"Mysqlx.Sql.StmtExecute"> & {
  /**
   * * namespace of the statement to be executed
   *
   * @generated from field: optional string namespace = 3 [default = "sql"];
   */
  namespace: string;

  /**
   * * statement that shall be executed
   *
   * @generated from field: required bytes stmt = 1;
   */
  stmt: Uint8Array;

  /**
   * * values for wildcard replacements
   *
   * @generated from field: repeated Mysqlx.Datatypes.Any args = 2;
   */
  args: Any[];

  /**
   * * send only type information for @ref Mysqlx::Resultset::ColumnMetaData,
   * skipping names and others
   *
   * @generated from field: optional bool compact_metadata = 4 [default = false];
   */
  compactMetadata: boolean;
};

/**
 * Describes the message Mysqlx.Sql.StmtExecute.
 * Use `create(StmtExecuteSchema)` to create a new message.
 */
export const StmtExecuteSchema: GenMessage<StmtExecute> /*@__PURE__*/ = messageDesc(
  file_mysqlx_sql,
  0,
);